test: fix flaky box.cfg() env vars test
The test starts a child tarantool instance in the current working
directory and run box.cfg(). The current working directory is a source
directory. test-run.py runs many tests from it in parallel and in some
circumstances it appears that the directory is locked by some other
box.cfg() call from some other tarantool instance (maybe another test is
doing the same).

The test is rewritten to use a temporary directory for such runs.

The test.interactive_tarantool helper is used for convenience instead of
a popen wrapper that is written specifically for the test.

NO_DOC=It is a fix of a test.
NO_CHANGELOG=see NO_DOC

(cherry picked from commit dfa67ae63afc4487b0f39a02679e470d16994344)