xrow: add pointer to data in xrow_header
Object xrow_header is a parsed MsgPack header of xrow packet. Since we
are going to pass header and body of request in recovery triggers, we
need to pass a pointer to the header of binary packet. Let's add such
pointer to xrow_header object.

Part of #8859

NO_CHANGELOG=internal
NO_DOC=internal