cli: add --failover option and failover module stubs
The option enables a built-in failover coordinator script, which reads a
configuration from a file (pointed by --config option) or from etcd,
connects to all the instances, polls them for their status and performs
automatic actions to ensure that there is an active leader in each
replicaset (if possible).

The failover coordinator code is part of Tarantool Enterprise Edition.

Part of https://github.com/tarantool/tarantool-ee/issues/564

NO_DOC=The documentation request is to be added as part of Tarantool
       Enterprise Edition patchset.
NO_CHANGELOG=see NO_DOC
NO_TEST=To be tested in Tarantool Enterprise Edition.