static-build: return CMAKE_FIND_USE_CMAKE_SYSTEM_PATH=FALSE back
This option controls the default behavior of the following commands for
whether or not to search paths provided by platform-specific cmake
variables:
 - find_program()
 - find_library()
 - find_file()
 - find_path()
 - find_package()
By default this variable is not set, which is equivalent to it having a
value of TRUE.

-DCMAKE_FIND_USE_CMAKE_SYSTEM_PATH=FALSE was initially added by commit
800e5ed617f7 ("build: refactor static build process"), but it was later
removed by commit 7dc9fe44a5ae ("libunwind: add new submodule"). It seems
that we can return it back.

Closes #6998

NO_DOC=build
NO_TEST=build
NO_CHANGELOG=build