config: allow postponing object grants till creation
Before this patch privileges for space, function and sequence alone
were disabled. Only permission for universe could be granted. It was
done because there wasn't a way to grant or revoke a privilege when
subjecting object hasn't been created yet.

This patch allows grants and revokes for objects that will be created
after the applier. These actions will be postponed and executed via
on_commit trigger after the object is registered in the according system
space of box.space._space/_func/_sequence.

Now permission can be granted to space/function/sequence from within the
config.

Part of #8967

NO_DOC=documentation request will be filed manually for the whole
       credentials