core: refactor `port_dump_lua` operation modes
Currently, `port_dump_lua` has 2 modes of operation: 'flat' and 'tabular',
which are controlled via a boolean flag. We plan to extend it with another
mode of operation, 'Msgpack object', so we need to replace the flag with a
enumeration. Plus, this refactoring should improve readability.

Needed for #8147

NO_CHANGELOG=refactoring
NO_DOC=refactoring
NO_TEST=refactoring