downgrade: fix test to downgrade on available version
It is not convenient that test_downgrade_from_more_recent_version breaks
if we create tag for new version and do not add next version to the
downgrade versions list. If the version is released we should add it to
the list anyway but it is not matter of this test.

Follow up #9182

NO_DOC=internal
NO_CHANGELOG=internal