sql: introduce sql_space_by_src()
This patch introduces the sql_space_by_src() function. This function is
used to search for a space based on a given element of the
struct SrcList. The main purpose of this function is to introduce a
second lookup for spaces, so this function is not used where a second
lookup is not required.

Needed for #4467

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring