sql: second lookup for constraint names
This patch add second lookup for constraint name in the DROP CONSTRAINT
statement.

Part of #4467

@TarantoolBot document
Title: Second lookup for names in SQL

Since all names are now case sensitive, a second lookup is now supported
to maintain backward compatibility. A second lookup by name will be
performed if the object was not found by name in the first lookup and if
the name is not enclosed in double quotation marks. The name for the
second lookup is normalized using old rules.