space: drop space with its triggers set with old API
Attaching triggers to space id instead of space object is a significant
pitfall. The users who haven't discovered new triggers may not expect
the triggers of a dropped space will be fired by a new one. So let's
drop triggers that were set with old API along with the space.

All the tests, changed because of described above breaking change, are
restored.

Closes #9223

NO_DOC=later