test: increase expected selectG execution time for debug ASAN build
The test is quite a flacky in debug ASAN CI workflow. The issue is test
check upper boundary of it's execution time. I run many instances of
this test on in parallel and got average time of 40s for memtex and 70s
for vinyl.

The time quota is already changed by the commit 84cb1e049272 ("sql:
increase time quota for selectG test on vinyl") for laptops with HDD.
I did not check execution time for HDD though. I guess the bottleneck
for debug ASAN is CPU.

Follow-up #7327

NO_CHANGELOG=internal
NO_DOC=internal