box: fix memory leak when dropping temporary spaces
Before this fix we used to leak memory when dropping fully-temporary
spaces. The problem was that the memory is only freed in the on_commit
callback of the transaction, but dropping a temporary space is a NOP
(meta-data is not persisted), and the on_commit triggers used to not be
invoked for NOP transactions.

Closes #9296

NO_DOC=bug fix

(cherry picked from commit a90819bd83dedd2bf0299f30cadcbc05b2f89c43)