box: expose box.schema before box.cfg
box.schema has a number of constants, e.g. IDs of system spaces,
which may be useful for the user. Let's allow to access box.schema
before box.cfg is called.

It is used in checking names, as we need to know, which snapshot spaces
to scan.

Needed for #8978

NO_DOC=minor change